+++ src/main/java/com/epam/www/dto/EventDTO.java
@@ -1,13 +1,11 @@
 package com.epam.www.dto;
 
 import com.epam.www.dataaccess.entity.Event;
-import com.sun.media.sound.AiffFileReader;
 import org.hibernate.validator.constraints.NotEmpty;
 
 import javax.validation.constraints.Digits;
 import javax.validation.constraints.Min;
-import java.time.LocalDate;
-import java.time.LocalTime;
+import java.math.BigDecimal;
 
 /**
  * Created by Farkas on 2017.02.23..
@@ -18,7 +16,7 @@ public class EventDTO {
     @NotEmpty(message = "Title must not be empty!")
     private String title = "";
     @Min(value = 0, message = "Price cannot be negative!")
-    private long price;
+    private BigDecimal price;
     @Digits(integer = 10, fraction = 0 ,message = "Must be epoch time representation!")
     private long airDate;
     @NotEmpty(message = "Auditorium must not be empty!")
@@ -54,11 +52,11 @@ public class EventDTO {
         this.title = title;
     }
 
-    public long getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(long price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
